Stop z-motd from reporting failure when .updates doesn't exist

+++ z-motd.sh
@@ -3,4 +3,6 @@
 # Place in /etc/portage/postsync.d and drop the .sh suffix.
 # Would be nice if this could also run after every package installation.
 
-rm /var/tmp/.updates
+if [ -f /var/tmp/.updates ]; then
+    rm /var/tmp/.updates
+fi
